+++ src/hviske/experiment_tracking/__init__.py
@@ -2,3 +2,4 @@
 
 from .extracking_factory import load_extracking_setup
 from .extracking_setup import ExTrackingSetup
+from .wandb_setup import preflight_wandb_access